requirements/base.txt
django-parsley==0.6
django-recaptcha==1.0.4
mysqlclient==1.3.7
python-magic==0.4.10
Unipath==1.1
wheel==0.24.0
zipstream==1.1.3
